requirements.txt
Flask==2.2.5
gunicorn==21.2.0
boto3==1.34.82
openai>=1.0.0
PyMuPDF==1.24.9
mutagen==1.47.0
google-generativeai==0.5.0
Pillow==10.3.0